Deleted The Food Object
Had to delete the food object and all food.property variables due to lack of knowledge. Going to restart from scratch with the food (red rectangle) object so it is a multiple of 20 because the snake moves by intervals of 20 and I want snake.x to equal food.x to the dot.